exchangeを抽出する際の計算にfloatを使うように変更。誤差が減る

+++ domain/debt.go
@@ -1,14 +1,15 @@
 package domain
 
 type Debt struct {
-	Price  int
+	// 割り算が絡むのでfloatにして後で均す
+	Price  float64
 	Debtor *User
 }
 
 type DebtCollection []Debt
 
-func (dc DebtCollection) DebtSum() int {
-	sum := 0
+func (dc DebtCollection) DebtSum() float64 {
+	sum := 0.0
 	for _, v := range dc {
 		sum += v.Price
 	}
